rename index.ts to main.ts and update imports
