+++ bin/minilogo.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
-import { default as main } from "../out/cli/index.js";
+import { default as main } from "../out/cli/main.js";
 main();
